Over draft protection

See merge request campuses/nlr/may-2023/java-purple/student-pairs/module-2-capstone-ws/nlr-14-module-2-capstone-team-6!39

+++ tenmo-server/src/main/java/com/techelevator/tenmo/controller/AccountController.java
@@ -24,8 +24,8 @@ public class AccountController {
     }
 
     @RequestMapping(value = "/balance", method = RequestMethod.GET)
-    public double getBalance(Principal p) {
-        double balance = 0.0;
+    public BigDecimal getBalance(Principal p) {
+        BigDecimal balance = 0.0;
         int userId = userDao.findIdByUsername(p.getName());
         Account account = accountDao.getAccountByUserId(userId);
         balance = account.getBalance();
